Remade database layer with Dagger. Created database Mock

+++ app/src/main/java/com/z80h3x/kezd_kov/ui/char_details/CharDetailsViewModel.kt
@@ -7,7 +7,7 @@ class CharDetailsViewModel @Inject constructor(
     private val charDetailsPresenter: CharDetailsPresenter
 ) : RainbowCakeViewModel<CharDetailsViewState>(Loading) {
 
-    fun load(id: Int) = execute {
+    fun load(id: Long) = execute {
         viewState = CharDetailsReady(charDetailsPresenter.getCharacter(id))
     }
